stream-trigger-job/requirements.txt
oci==2.44.1
requests
python-dotenv